# swagger/doc.yml
swagger: "2.0"
info:
  description: "A GinBro RESTful APIs"
  version: "1.0.0"
  title: "GinBro RESTful APIs Application"
host: "127.0.0.1:8082"
basePath: "/api/v1"
schemes:
- "http"
paths:
...
  /cart/{Id}:
    get:
      tags:
      - "cart"
      summary: "get a cart by ID"
      description: "get a cart by ID"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of cart to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Cart"
    delete:
      tags:
      - "cart"
      summary: "Destroy a cart by ID"
      description: "delete a cart by ID"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of cart to delete"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
...
  /product:
    get:
      tags:
      - "product"
      summary: "get all product by pagination"
      description: ""
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "where"
        in: "query"
        description: "column:value will use sql LIKE for search eg:id:67 will where id >67 eg2: name:eric => where name LIKE '%eric%'"
        required: false
        type: "array"
        items:
          type: "string"
      - name: "fields"
        in: "query"
        description: "{$tableColumn},{$tableColumn}... "
        required: false
        type: "string"
      - name: "order"
        in: "query"
        description: "eg: id desc, name desc"
        required: false
        type: "string"
      - name: "offset"
        in: "query"
        description: "sql offset eg: 10"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        default: "2"
        description: "limit returning object count"
        required: false
        type: "integer"

      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ProductPagination"
    post:
      tags:
      - "product"
      summary: "create product"
      description: "create product"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create product"
        required: true
        schema:
          $ref: "#/definitions/Product"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"

    patch:
      tags:
      - "product"
      summary: "update product"
      description: "update product"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create product"
        required: true
        schema:
          $ref: "#/definitions/Product"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
  /product/{Id}:
    get:
      tags:
      - "product"
      summary: "get a product by ID"
      description: "get a product by ID"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of product to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Product"
    delete:
      tags:
      - "product"
      summary: "Destroy a product by ID"
      description: "delete a product by ID"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of product to delete"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
  /user:
    get:
      tags:
      - "user"
      summary: "get all user by pagination"
      description: ""
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "where"
        in: "query"
        description: "column:value will use sql LIKE for search eg:id:67 will where id >67 eg2: name:eric => where name LIKE '%eric%'"
        required: false
        type: "array"
        items:
          type: "string"
      - name: "fields"
        in: "query"
        description: "{$tableColumn},{$tableColumn}... "
        required: false
        type: "string"
      - name: "order"
        in: "query"
        description: "eg: id desc, name desc"
        required: false
        type: "string"
      - name: "offset"
        in: "query"
        description: "sql offset eg: 10"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        default: "2"
        description: "limit returning object count"
        required: false
        type: "integer"

      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/UserPagination"
    post:
      tags:
      - "user"
      summary: "create user"
      description: "create user"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create user"
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"

    patch:
      tags:
      - "user"
      summary: "update user"
      description: "update user"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create user"
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
  /user/{Id}:
    get:
      tags:
      - "user"
      summary: "get a user by ID"
      description: "get a user by ID"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of user to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/User"
    delete:
      tags:
      - "user"
      summary: "Destroy a user by ID"
      description: "delete a user by ID"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of user to delete"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
...
definitions:
...
  Cart:
    type: "object"
    properties:
    
      id:
        type: "integer"
        description: "主键"
        format: "int64"
      
      create_time:
        type: "string"
        description: "创建时间"
        format: "date-time"
      
      update_time:
        type: "string"
        description: "更新时间"
        format: "date-time"
      
      create_user:
        type: "integer"
        description: "创建人"
        format: "int64"
      
      update_user:
        type: "integer"
        description: "更新人"
        format: "int64"
      
      delete_flag:
        type: "integer"
        description: "删除标志"
        format: "int64"
      
      user_id:
        type: "integer"
        description: "用户id"
        format: "int64"
      
      activity_id:
        type: "integer"
        description: "活动id"
        format: "int64"
      
      activity_name:
        type: "string"
        description: "活动名称"
        format: "string"
      
      product_id:
        type: "integer"
        description: "商品id"
        format: "int64"
      
      product_name:
        type: "string"
        description: "商品名称"
        format: "string"
      
      product_subtitle:
        type: "string"
        description: "商品简要描述"
        format: "string"
      
      product_main_image:
        type: "string"
        description: "商品图片地址"
        format: "string"
      
      quantity:
        type: "integer"
        description: "数量"
        format: "int32"
      
      product_unit_price:
        type: "number"
        description: "单价"
        format: "float"
      
      selected:
        type: "integer"
        description: "是否已选择 1是 0 否"
        format: "int64"
      
      product_total_price:
        type: "number"
        description: "总价格"
        format: "float"
...
  Product:
    type: "object"
    properties:
    
      id:
        type: "integer"
        description: "主键"
        format: "int64"
      
      create_time:
        type: "string"
        description: "创建时间"
        format: "date-time"
      
      update_time:
        type: "string"
        description: "更新时间"
        format: "date-time"
      
      create_user:
        type: "integer"
        description: "创建人"
        format: "int64"
      
      update_user:
        type: "integer"
        description: "更新人"
        format: "int64"
      
      delete_flag:
        type: "integer"
        description: "删除标志"
        format: "int64"
      
      product_id:
        type: "integer"
        description: "商品id"
        format: "int64"
      
      category_id:
        type: "integer"
        description: "品类id"
        format: "int64"
      
      name:
        type: "string"
        description: "商品名称"
        format: "string"
      
      sub_title:
        type: "string"
        description: "简要描述"
        format: "string"
      
      main_image:
        type: "string"
        description: "商品图片地址"
        format: "string"
      
      sub_images:
        type: "string"
        description: "子图片列表"
        format: "string"
      
      activity_id:
        type: "integer"
        description: "活动id"
        format: "int64"
      
      status:
        type: "integer"
        description: "商品状态"
        format: "int64"
      
      price:
        type: "number"
        description: "商品单价"
        format: "float"
      
      stock:
        type: "integer"
        description: "库存数"
        format: "int64"
  ProductPagination:
    type: "object"
    properties:
      code:
        type: "integer"
        description: "json repose code"
        format: "int32"
      total:
        type: "integer"
        description: "total numbers"
        format: "int32"
      offset:
        type: "integer"
        description: "offset"
        format: "int32"
      limit:
        type: "integer"
        description: "limit"
        format: "int32"
      list:
        type: "array"
        items:
          $ref: "#/definitions/Product"
  User:
    type: "object"
    properties:
    
      id:
        type: "integer"
        description: "主键"
        format: "int64"
      
      create_time:
        type: "string"
        description: "创建时间"
        format: "date-time"
      
      update_time:
        type: "string"
        description: "更新时间"
        format: "date-time"
      
      create_user:
        type: "integer"
        description: "创建人"
        format: "int64"
      
      update_user:
        type: "integer"
        description: "更新人"
        format: "int64"
      
      delete_flag:
        type: "integer"
        description: "删除标志"
        format: "int64"
      
      status:
        type: "integer"
        description: "启用标志"
        format: "int64"
      
      user_name:
        type: "string"
        description: "用户名"
        format: "string"
      
      email:
        type: "string"
        description: "邮箱"
        format: "string"
      
      phone:
        type: "string"
        description: "手机号"
        format: "string"
      
      password:
        type: "string"
        description: "密码"
        format: "string"
  UserPagination:
    type: "object"
    properties:
      code:
        type: "integer"
        description: "json repose code"
        format: "int32"
      total:
        type: "integer"
        description: "total numbers"
        format: "int32"
      offset:
        type: "integer"
        description: "offset"
        format: "int32"
      limit:
        type: "integer"
        description: "limit"
        format: "int32"
      list:
        type: "array"
        items:
          $ref: "#/definitions/User"
...
  ApiResponse:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      msg:
        type: "string"
...